requirements.txt
matplotlib==3.7.1
numpy==1.24.2
pandas==2.0.0
Pillow==9.5.0
prettytable==3.7.0
torch==2.0.0
torchinfo==1.7.2
torchvision==0.15.1
tqdm==4.65.0
opencv-python
tensorboard